clips-executive: re-factor CX initialization
The initialization is now fully controlled by the configuration. It is
simplified and more generalized. There is no code specific to a
particular feature (aside configuring debug information) anymore. This
is all handled generically. We specify a number of stages, but the
stages are fully controlled through the configuration giving maximum
flexibility.