robot-memory: fix triggers for shared database
The initialization of triggers for the shared database was broken. It
compared the namespace always with the name of the replica set client
config, two unrelated things. This lead to triggers being created on the
local database always. Triggers on the distributed database were
virtually impossible with this.

This explains the failure to use robot-memory last year in the
clips-agent. Syncing data back could not have worked.