clips-robot-memory: provide explicit feedback on try-lock fail
Try-lock may fail for an anticipated reason: the lock has already been
acquired. In this case, however, the lock fact is not updated and thus
this cannot be caught by a trigger. This is an issue in particular if it
occurs for a mutex which was previously unknown (e.g., after a hot start
with an existing DB or when joining a running replica set).

The solution is to take the result into account and assert an indicator
fact to the calling environment indicating the error.