clips-executive: cancel the Skiller if its action disappeared
Send a StopExecMessage to the Skiller if no corresponding plan action
exists, or if that action was retracted. This can happen if we cancel a
goal or an action.

This fixes an issue where the agent would not execute any other skill if
a previous action was failed or canceled while the skill was still
running.