clips-executive: do not convert data from robot memory a second time
The functions bson-get-array and bson-get already return the proper
type, do not try to cast from string into the type again.

This fixes an issue with syncing non-string values, which would result
in errors because the function eval expects a string or a symbol.

Thanks to Tarik Viehmann for reporting and debugging this.