buildsys: drop support for generated interface code in repo
So far, we supported systems that did not have the pre-requisites for
ffifacegen, i.e., we allowed to re-use interfaces that had been
generated before and committed to the source repository.

This has several drawbacks:
- we virtually (possibly actually) never used the pre-generated files.
  libxml2 is readily available on all target systems.
- repository is inflated by code which can be and usually is easily
  generated from the description files
- In some constellations code would fail to build
- Updates to interfaces are error prone

Since the only benefit (allow libxml2 to be missing) is of virtually no
relevance, drop support for this feature.

This currently has a small caveat in that it breaks system installation.
This can be fixed later with some minor tweaks to the installation build
type.