buildsys: add format-emacs target
This calls Emacs in batch mode to format files. Note, however, that this
does not produce the exact same style as clang-format. It is informative
and to enable tuning of the Emacs style to more closely match the
clang-format mandated style (which is authoritative).

This contains a minimal Emacs configuration to configure code
formatting. Some more effort is necessary to more closely align Emacs
and clang-format.

Using Emacs is an order of magnitude slower than clang-format.