libplugin: allow meta plugins as YAML lists
Defining a meta-plugin as a one-line string has been a long-standing
pain in every merge-meisters brain. Feature branches have a tendency to
change meta-plugins, and if all plugins are listed in one line that
causes unnecessary conflicts.

With this commit, meta-plugins can be defined via YAML lists, like so:

some-meta-plugin:
  - real-plugin-1
  - real-plugin-2

This eliminates the tedious (and often repetitive) conflict resolution
when merging long-running feature branches. The old syntax is of course
still supported (but strongly discouraged).