buildsys: add implicit parallel-* targets
This is intended for automatic builds for continuous integration. You
probably should not use this on the command line. Note, that this may be
stacked with uncolored, but "parallel-" must be the prefix, e.g.,
parallel-uncolored-all.

Note that the parallel target is a root target. It only works in the
top-level directory, not for individual builds for, e.g., plugins.