clips-executive: make REJECTED an outcome instead of a mode
We also need to evaluate the goal if it was REJECTED, e.g., to release
acquired resources. Also, REJECTED is similar to FAILED, except that a
goal can not be rejected if it has already been COMMITTED or DISPATCHED.
For this purpose, remove the REJECTED mode and add REJECTED as an
outcome instead.