clips-executive: add required and acquired resources to the goal spec
A required resource is a resource that the agent needs to hold before
the goal can be dispatched. An acquired resource is a resource that the
agent is currently holding. A goal can only be dispatched if the
required resources are a subset of the acquired resources. A resource is
NOT removed from the required resources when the agent gets the
resource.