clips-executive: add lock/unlock mechanism for resource locks
This automatically locks and unlocks the resources required by a goal.
If the goal is COMMITTED and a resource has not been acquired yet,
try to lock the mutex with the name of the resource.

If the lock is acquired, add the resource to the acquired resources. If
the lock is rejected, reject the goal.

Once a goal is REJECTED, unlock all its resources.