clips-executive: directly reject a goal if a resource is locked
We often know the states of the mutexes, no need to always try to lock
it. Instead, directly reject the goal.

This also fixes an issue where a second goal is committed and then tries
to get a mutex that is already locked by the same agent because of a
different goal, thereby leading to a mutex error.