clips-executive: on wm update, only delete domain objs of the same type
If an object is removed from the worldmodel (by removing it of the list
of objects of that type), only delete those domain objects that have the
correct type.

This fixes a bug that would delete all domain objects of a different
type if one object is removed from the worldmodel.