clips-executive: wait for sub-goals to be evaluated
For sub-goal sub-types, wait for sub-goals to be in EVALUATED mode
before determining the parent goal's fate. This enables to reason about
the outcome of sub-goals before continuing.

This requires that all sub-goals eventually end up in the EVALUATED
state.

Thanks to Mostafa Gomaa for a helpful discussion of the sub-goal
sub-types with respect to the RCLL domain implementation.