clips-executive: truncated systime for update timestamps in robmem sync
When running in simulation, timestamp may diverge between system and in
Fawkes (using simulated time). However, deleting facts on a robot memory
triggered update relies on the timestamp provided by MongoDB, and this
is using system time.

Use the new (now-systime) function to use system time for remembering
update timestamps. Additionally, truncate timestamps to milliseconds
precision, as this is the precision that MongoDB can handle internally.

Thanks to Till Hofmann for joint debugging of the issue.