clips-executive: support automatic cleanup of RETRACTED goals
For goals which are retracted, always cleanup associated plans. For
RETRACTED goals we make no promises, or rather, the promise you cannot
rely on it anymore. So it is safe to immediately remove the plan.

For RETRACTED goals, cleanup the goal tree iff the goal has no parent or
the parent is a MAINTAIN goal. The reasoning is that we want to keep
RETRACTED goals which are part of a goal tree to understand the choices
and run-time development  while the tree is still active. The cut-off
condition to identify the root of a tree is if it has no parent, or the
parent is a MAINTAIN goal which can never be RETRACTED and which may
ever only have a single ACHIEVE sub-goal.

Automatic goal cleanup is configurable through the new global
automatic-goal-retraction flag.