clips-executive: make sure there is no acquired resource when rejecting
We always need to release all acquired resources before rejecting the
goals. We take care of that by rejecting a goal only if the goal's
acquired resources slot is empty. However, there may still be a mutex
that was acquired without being processed (and added to the list of
acquired resources) yet. Thus, also check that there is no mutex that is
part of the required resourced and that has been acquired.