ros-navigator: generalize trans vel for dynreconf
The navigator_thread assumed to have two dynamically configurable
parameters available in the local planner used in move_base - one
for the x-velocity and one for the y-velocity. However, the
NavigatorInterface provides a SetMaxVelocityMessage to set only
one translational speed. We can not know whether the local planner
in move_base actually allows to set a second translational speed.

To adapt towards the interface's SetMaxVelocityMessage we only set
one translational velocity as a dynamically reconfigurable parameter.