clips-executive: only make mode decisions in goal reasoner
To clarify responsibilities of the component, it is only the goal
reasoner which can switch goal modes now. The new plan therefore no
longer modifies the goal. Rather, the goal reasoner checks for selected
goals which have a valid plan and then marks them as expanded. Later,
more specific cases can account for ensemble planning with multiple
planners and plans.