plexil: add adapter for basic skill execution
It's still a little bit awkward on the Plexil side. But it may be
inherent to Plexil itself, i.e., that all actions are finished
immediately and an end condition is required to wait for the outcome.
It may have some benefits we yet have to discover. We may want to
implement a slightly different pattern to better support synchronous
command execution for skills.