plexil: support verbatim XML adapter arguments in config
Some adapters may require a rather complex XML-based configuration. For
example, the UdpAdapter expects message specifications. To enable this
use case, we now support a verbatim-xml field, which can be a partial
XML document. On initnialization, it is parsed (and therefore we ensure
that it indeed is valid XML) and all children are injected into the
adapter's interface config tag.