plexil: add automatic compilation support
The plugin can now invoke the plexil compiler automatically as necessary
(or forced every time). Since Plexil is not C++17 compatible at this
point, resort to using boost::filesystem.

Compilation, if enabled, triggers when the plx is missing or older than
the ple, or if compilation is forced. The plx name can be generated
automatically from the ple.

Running with an empty plan is no longer supported.