plexil: enable running in external exec using FawkesRemoteAdapter
The new FawkesRemoteAdapter can be used to run a plan externall in the
universal executive, rather than embedded into Fawkes.

It looses features such as auto-compilation, but may be handy to restart
runs quickly without always reloading the full executive/Fawkes
instance, especially when considering that the Plexil app framework is
not well-suited for dynamic reloads.