libconfig: fix YAML config value iterator for get_value
The value iterator was invalid, even when a valid value had been
retrieved during the generic Configuration::get_value() for the YAML
configuration. The reason is that it assumed that next would be called
before querying for validity. This is true for search results, but not
for value retrieval.

This is an uncommon use case when using the configuration directly.
However, in the case of the in-memory config used by the network
configuration, the ValueIterator::is_valid() would be called to check
for existence of a value, then always returning false.

Fix by properly initializing to the first value, and skipping an advance
on the first call to next().