plexil: enable automatic global state registration
If the GlobalState interface is configured as default lookup adapter,
states are added automatically when they are set for the first time.
Any non-existing global will yield UNKNOWN as value. Multiple set calls
with inconsistent value types will be detected and rejected.

This allows for a more flexible way to use the module without needing to
modify both, ple and yaml config files for the same state.

The configuration may still be used, which can be useful in particular
to enforce a specific value type, or to provide initial values.