buildkite: generalize SSH identity handling for in-script cloning
Previously, we had a special handling for the committers repository key,
assuming that we only clone this one repository within the build script
(not the checkout of the code repository). However, this fails for the
in-script sub-clone of a bare repository for author/committer linting.
For private repos, we need to use a proper SSH deploy key/identity.

Extend the scripts to generalize the extraction of keys from environment
variables. This is compatible with our expection of what the environment
agent hook is supposed to do.

Remove the specific committer key handling from the lint script.