openrave: uncomment setting of unused local variable
Looking closer into this reveals that nonlocal was not the issue.
Rather, it is used in a while loop that will always exit after the first
run (the return is on the same level).

At this point the intended behavior is not entirely clear. This commit
only fixes the warning keeping the same contract.