pantilt: fix return value handling in DP PTU
This is one of the more elaborate fixes necessary for lgtm findings.
There are two issues:
1. the read() method did not distinguish no bytes read from not enough
bytes read. Analyzing our code, we are fine with simply checking for
"some bytes up to the given maximum buffer size have been read".
2. The query_int() method assumed a read() syscall rather than the
method call. Therefore, even "false" returns would be accepted as
"everything ok". With the change of 1., we can now simply check if some
data has been read (we cannot know the overall size in advance) and then
leave it to sscanf() to verify the result.

Tested on the actual device and it's (still) working nicely.

Detected by https://lgtm.com