libconfig: add provision for input SQL check
We pass lines verbatim from the input file to sqlite3_exec(). In our
case this is ok, it is only used on a previously empty database to
import a dump. However, to avoid disabling this lgtm rule altogether
(cpp/sql-injection), add a function that could be used to verify the
input more closely. Detected by https://lgtm.com.