buildkite: add conditional step to build webview frontend
The frontend requires compilation and therefore is an own worthwhile
build target. However, it most certainly does not make sense to build it
every time. Therefore, use a dynamic step to build only for the master
branch, a branch which contains the substring "webview", or if the
affected files contain anything in the frontend directory.