laser: fix type error
This was introduced in 3765f31 when fixing a Codacy error. This was
raised for good reason: we are using sscanf() to read a "%x" format. The
documentation (man sscanf) states that this must be read into an
unsigned integer. The Sick TiM protocol documentation indeed states the
type "Uint_32" for the starting angle field, however, the range is
declared as -450000d to +2250000d.

So we need to read an unsigned int and then interpret it as a signed
integer to be compliant with sscanf and the device protocol.

Thanks to @vmatare for reporting this issue.