buildsys: also check SRCDIR for dynamic source files
If a source file is not available at Make dependency resolution, i.e.,
it is generated or copied only as part of the build process, then
VPATH-based path resolution will fail to reference the proper file. This
is an issue if the generate file does reside in SRCDIR, rather than
OBJDIR (as is the case, for example, for *_tolua.cpp files).

To remedy this, check if a file of the requested name exists in SRCDIR,
if it does, use that. Otherwise, use the previous resolution, which may
be VPATH-based resolution or verbatim copy, i.e., look in OBJDIR.