googletest: pseudo library to build gtest if necessary
On Ubuntu, annoyingly there is no pre-built libgtest (as there is on
Fedora or FreeBSD). Add a pseudo-library, which will, in that case, copy
the files from the system and build them with the Fawkes build system,
effectively using the exact same compiler flags. The libraries are named
fawkesgtest and fawkesgtest_main to avoid confusion with pre-build
libraries.