buildsys: make *_all declarations for maximum extent build
This fundamentally changes the meaning of variables such as LIBS_all,
BINS_all, and PLUGINS_all.

Up to now, it used to describe the targets to build in this very call
(targets not being a gui or a test target).

With this change, it is a declaration of the maximum extent of such
targets, i.e., it lists all targets that could be built if all
requirements were met. The new variables LIBS_build, BINS_build,
and PLUGINS_build denote which targets out of these should be built, as
do the already previously existing *_gui and *_test targets.

Adapt interface and protobuf default rules to this change generating the
appropriate LIBS_build targets.

This change is necessary in order to reliably detect skipped targets.
The alternative, to define *_skipped variables, would require Makefile
authors to appropriately set this variables in every case where a module
is not built. This could quickly lead to oversights and inconsistencies.
In order to help (especially with the transition), the build system will
complain whenever a target is marked to be built but has not been added
to the respective _all variable.