buildkite: parallelize gitlint
gitlint does not support internal parallelization, worse, it seems to be
particularly slow when called with the "--commits" option for the
branch's range.

Retrieve the log messages from git and feed each to a separate instance
of gitlint. Parallelize by spawning as many parallel processes as there
are (logical) CPU cores using xargs. The speedup is almost linear in the
number of cores.

Improve representation of the error annotation.