clips-executive: detect if action is no longer available
This if fixing the first issue contributing to #51:

The CX uses the PDDL-based planning model. There is an issue that can lead
in some (many or even all?) cases to actions being marked executable, when
they really are not. The problem is two-fold.

Once an action has been marked executable because its precondition is
fulfilled, it will never be re-evaluated. If the precondition is no longer
satisfied, the executability is not reset. Once executable, always
executable.

This change fixes this aspect by adding a rule that marks an action not
executable when the precondition is no longer satisfied.