clips-executive: check satisfiability of preconditions of if atoms exist
This if fixing the first issue contributing to #51:

The CX uses the PDDL-based planning model. There is an issue that can lead
in some (many or even all?) cases to actions being marked executable, when
they really are not. The problem is two-fold, this change tackles the
second issue:
When grounding the precondition of a plan-action, the domain-precondition
is asserted first, only then are the related domain-atomic-precondition
facts asserted. Inbetween, a conjunctive (or disjunctive for that matter)
precondition is deemed satisfied, as there are no contradicting parts
thereof. So the conjunction will always be satisfied for a brief moment.
Once the atomic preconditions are grounded (and not satisfied), the
conjunction will in turn be no longer satisfied. However, with the
earlier problem that executability was never reset, this is not
considered for execution.

This change makes additions to rules handling satisfaction of
conjunctive and disjunctive preconditions that ensure that for each
atomic precondition depending on the compound precondition in the
ungrounded model, a grounded instance also exists. This way, we can be
sure that all the parts to determine satisfiability are really present
at the time of making that determination.

This, combined with the first part, is a serious bug, as it has the
potential to bypass many, if not all, precondition checks altogether.
Once an action is marked executable it is never re-checked, and the
second part means that for compounds we decide this too early and then
always assume it has been satisfied. It depends on the order of rule
activations, but we have to assume that this was essentially broken all
the time. I expect for existing domain models and existing (fixed or
template) plans to break or get stuck. The upcoming work on execution
monitoring can at least raise the necessary red flags to detect and fix
these problems.