webview-frontend: fix endless loop when updating CX goal detail
When refreshing the goal detail view and the goal was RETRACTED (and
auto cleanup of retracted goals is disabled, i.e., the goal stays
potentialy indefinitely) a loop would occur resulting in endless
refreshing of the goals, eventually crashing the browser.

Fix by accounting for the RETRACTED goal mode for the refresh decision.