webview-frontend: build only using local Angular CLI
With this change, we no longer use a globally installed Angular CLI
version, but only the local version that comes with "npm install" or
"npm ci" is required. The build flags are now properly maintained in
package.json (earlier commit) and have been removed from the Makefile.
It seems more appropriate to configure them there.