syncpoint: use barrier in test for startup sync
Eliminate a fragile startup synchronization with a barrier. However,
this is only an intermediate effort on the wait to the more permanent
solution we just discussed.

This already mitigates many test fails on my system. However, as
discussed, there are still looming race conditions and a more robust fix
is possible.