buildkite: account for the case that there are no topic branches
Linting (on the master) would fail if there were no topic branches. We
have had so many and persistent topic branches that this is the first
time this bug actually triggers. Two places are problematic: due to "set
-e" and grep returning 1 for "no lines were selected" (cf. man grep).
Then even if this is fixed, the rev-parse will fail since "--not"
expects an argument. Hence, make this contingent on having any other
branches at all.