buildsys: fix formatting check script for broken clang-format
Running a recent clang-format (tested with 7.0.1) yields replacements in
the XML output even though the file is properly formatted and another
clang-format run would not change it.

Work around by comparing the original and a reformatted file. The effort
is marginally higher. The old version also ran a full clang-format, but
could avoid the full diff. But it's still fast.