clang-format: do not allow short functions or if statements on same line
Having functions on the same line is confusing. Code can be missed,
because it is at the end of a line, especially when in a header file it
can be easily missed. It also destroys the similarity among functions.

The same holds for if statements, especially since they would otherwise
be inconsistent with functions.