buildsys: copy format git tree filter script out-of-tree for format-branch
If using a script for the tree filter which is under version control
itself, the git tree filter would potentially older versions, if the
script itself had been modified in the branch in question.

Avoid this by explicitly checking out the latest version for each commit
to an out-of-tree version and then deleting it again.