buildsys: add local format target
So far, there was a format target in the root directory, which reformats
everything. With this change, there is a new local format target, which
is triggered when running "make format" in a source directory. It will
format all locally modified and untracked files in the given directory.

This comes very handy to quickly format the code before starting
committing.