buildkite: fix builds from fork repos
When issuing a build for an external fork repo (typically doing a
rebuild on a previously rejected build) using BUILDKITE_BRANCH yields
something like "user:branch" (we explicitly enabled that in buildkite
and use it for, e.g., branch name verification). This will lead the
author information check to fail since it would use a commit range like
"MERGE_BASE..user:branch". Fix this by using BUILDKITE_COMMIT. Make sure
to add this to the environment passed to the docker containers.