laser-filter: fix angle calculation in projection filter
The projection filter needs to determine the angle of an indexed beam.
For input data of 360 and 720 entries it uses specialized lookup tables.
However, for the case of 1080 entries it is determined by a formula.
This formula, however, is incorrect.

Fix the calculation by using the same formula underlying the generation
of the lookup tables.