git-hooks: add pre-push hook that enforces checking before pushing
Our growing rule set and lack of pre-checks by committers leads to a
high number of failed builds.

To improve this a little bit and train people to run "make check" before
pushing soft-enforce this by installing a local pre-push hook on the
first run of make. This then checks a timestamp file created/touched by
"make check" to check if any of the files affected on the current branch
has been modified after running the check last.

This is not bullet proof, if someone develops on the local master branch
and pushes with something like "git push master:some/branch" the check
would not work. However, this is then still caught by the CI builds.