mongodb: fix initialization of ReplicaSetConfigThread
The ReplicaSetConfigThread was used by the MongoDBThread before it was
fully initialized, resulting in segfaults, e.g., by logger being NULL.

Fix the initialization in two ways: For one, move most of the
ReplicaSetConfig initialization from the constructor into init().
Second, do not call bootstrap from the mongodb thread, but instead call
it at the end of init. This way, we can guarantee that the
ReplicaSetConfigThread is fully initialized before operating on it.