robot-memory: switch from mongodb oplog to change streams
Use change streams instead of querying the oplog for all event triggers.
Always return the full document if an event is triggered.

Currently, we only support empty queries for the event trigger, i.e.,
the trigger will fire for any change in the given collection. If the
given query is non-empty, throw an exception.

This requires further adaptions in event trigger consumers, as the
resulting document's structure has changed.