clips-executive: use current time as timestamp for syncing fact deletion
Instead of using the fact creation time, use the current time when
syncing a fact deletion from the local world model to robot memory. This
avoids an issue where the deletion would not be synced to another robot
if the other robot created the same fact but with a later timestamp.

We do not know exactly when the local wm-fact was retracted, so we
cannot use the exact retraction timestamp. However, the current time
will be almost correct, because the rule fires immediately when the fact
is retracted.

This fixes carologistics/fawkes-robotino#27.