clips-executive: change lock response field directly
on lock renewal, the response of the mutex would be changed from
pending to acquired or rejected upon the receival of the mutex-op-feedback
fact. The decision when the next lock renewal have to be started is
based on the lock-time of the mutex. This would have been updated only
if the robmem-trigger of the renewal arrived before the mutex-op-feedback.
If this was not the case, the renewal started again instantly

In order to prohibit this race, we have to adapt the response field
directly when receiving the robmem-trigger, instead of a seperate rule. To make
this consistent with normal locking, this is changed generally for all requests