robot-memory: fix query function
The query function still used the old way of getting the mondodb
client needed. Get the collection directly instead of using the client,
as in the other robot-memory functions.

Additionaly the read_preference was always set to secondary. However, if
this is not always the case. Therefore, use the passed query options instead
of overwriting the read_preference.