clips-executive: also check pending-requests for unlock requests
By the introduction of lock renewals, mutexes can have
multiple pending-requests. While unlocking resource locks,
mutexes were checked if they already have an unlock pending,
but only the request field was checked.
Additionally, when checking if an auto-renew is running,
the pending-requests field was matched with AUTO-RENEW-PROC,
which fails if there is any additional request pending.
Thus, this has to change to a member check of AUTO-RENEW-PROC
in pending-requests.