clips-executive: deal with oldest robmem-trigger first
Sometimes it can happen, that multiple robmem-trigger are asserted
in one step. This is no problem, until 2 or more of them are from
the same mutex or fact. Unfortunately clips handles the newest trigger
first. So if a mutex is unlocked and then locked again, in clips it
will get locked and then unlocked.

To deal with it, enforce clips to deal with the oldest robmem-trigger
first