robot-memory: catch exception if we fail to re-create a change stream
If there is currently no primary in the replica set (e.g., because it
died and leader election has not been triggered yet), the current change
stream fails and re-creating a change stream also fails. In that case,
catch the exception and keep trying in the next iteration, until a new
primary has been elected.

This also means that no trigger events will occur as long as there is no
new primary.