realsense: restart camera if polling fails too often
Sometimes the device has an USB error (e.g., due to insufficient power
supply). In that case, the device will not recover and needs to be
restarted. Count the number of unsuccessful polls and restart the camera
if that reaches a threshold.

The threshold is configurable via 'restart_after_num_errors' and
defaults to 10 tries.