realsense: check every loop if the camera needs to be started
We previously only checked if we need to start the camera if we were
using the switch. However, we may also need to start it because it
previously failed and needs to be restarted. Always check the current
and desired status of the camera and start/stop the camera if they
differ.