clips: properly deal with bool arrays
If the type of blackboard field is a boolean, the blackboard feature
treated the field as a single value field. However, this is not granted.
E.g. the RobotinoSensorInterface uses a bool array field for publishing
the states of the digital inputs.

Thus, treat bool fields similar to float or double fields and
replace false with FALSE and true with TRUE to match CLIPS boolean
representation.