clips-robot-memory: make register trigger asynchronous
There is a potential deadlock between the clips environment and
robot-memory if we do the trigger regisration synchronously (#90).

To avoid the deadlock, make the register_trigger function asynchronous:
It creates a future that will assert a feedback fact into the clips
environment once registering the trigger has succeeded or failed. This
is similar to the pattern used for mutex_try_lock_async.