clips-executive: keep robmem sync map entries of deleted facts
If we delete a fact, we should keep the sync map entry so we know when
we deleted the fact. This is necessary if we get an update for the same
ID, in which case we need to compare timestamps to decide whether we
should re-add the fact. To do this, set the wm-fact-idx to 0 on fact
deletion and also always update time update-timestamp, both when we
delete the fact locally and when we get a sync-delete from robot memory.

Co-Authored-By: Tarik Viehmann <tarik.viehmann@rwth-aachen.de>