ffifacegen: Correct msg name comparison in create_message
In the BlackboardNetInterfaceHandler interface messages are serialized
to send them over network to other nodes.
During these procedure, the message name is truncated to a length of
INTERFACE_MESSAGE_TYPE_SIZE_ - 1 characters, with a null character at
the INTERFACE_MESSAGE_TYPE_SIZE_'s position.
However, when the network message is received at the other node a new
message is created by the create_message function of the interface.
This function compares the message stored in the network message with
possible message names inside its interface. As this comparison happens
over max INTERFACE_MESSAGE_TYPE_SIZE_ characters the null character is
included and hence every comparison fails! (As long as the message name
is long enough.)

This commit correct the number of characters which are compared inside
the create_message function.