buildsys: only install targets that have been built
We changed the meaning of BINS_all, PLUGINS_all, .... They used to
denote the targets that are actually built. Now, a target is always
added to the respective _all variable, even if it is not built. Build
targets are now stored in BINS_build, .... Only install those, as the
other targets are not installable.