interfaces: make message copy constructor explicit
Forbid implicit copy construction of a message, as this may hide other
bugs which are caused by accidentally copying a message rather than
using a pointer to the message. In particular, this avoids implicit copy
construction when processing the message queue:

    Message m = iface->msgq_first<Message>();

This will (rightfully) no longer compile, because it implicitly
copy-constructs m from the Message* returned by msgq_first.