clips: fix underscores for preloaded blackboard interface name
Revert the name of the interface to `__clips_blackboard_preload__`. The
name does not really matter as the interface is immediately closed again
anyway, but this name more clearly marks the interface as an interface
not to be used. The interface name was changed erroneously in commit
dc44f15e47.