robot-memory: make robot memory thread a continuous thread
Decouple the thread from the main loop and let it run continuously.
Robot memory takes a long time to process trigger events, which blocks
the main loop and results in a lot of loop time warnings.

By default, use the main loop time as desired loop time and wait at the
end of the loop if the loop completed faster than desired. The loop time
can be overwritten in the config.

The disadvantage of running it continuously is that the triggers do not
run in the SENSOR_PROCESS hook anymore but may be triggered at any time.
However, as the thread takes too long and therefore is decoupled from
the main loop anyway, this is already happening now. Thus, this should
do no harm.