buildsys: expand globs in git ls-files with git, not the shell
If we call `git ls-files *.{h,cpp}`, then (depending on the shell), the
expansion of `*` may already be done by the shell instead of git. This
is undesired, as the shell does not list files in sub-directories, in
contrast to git. Thus, to make sure that we list all files, escape the
glob `*` so we always pass it to git. This issue occurs when using zsh,
but probably may also occur with bash, depending on the shell
configuration.