skiller: call states' reset() before clearing vars
A state's reset() method is a user-supplied method that is supposed to
perform custom cleanup actions on skill reset (e.g. when skill is
cancelled to execute another one). When implementing this method, a user
expects that the last state of the skill still exists, in particular
that the fsm.vars table hasn't been cleared, yet.