clips-executive: fix leaked resource in goal fast rejection
The rule is supposed to trigger only if there is no mutex with a pending
request, because otherwise we need to wait for the response for that
request. By re-using the variable name `?n`, which was bound before, we
only checked if there is no request for that particular mutex. Fix this
by using a different variable name.