clips-executive: keep track for which goal we requested a resource
If we have two goals that require the same resource, we need to keep
track for which goal we actually requested the resource. Otherwise, we
may request a resource for goal 1 and then assign it to goal 2. This is
undesirable, because goal 2 may require additional resources that we
have not requested, resulting in a deadlock.

On the way, split the rule for fast rejection into two rules, because we
can now directly check if another goal requested the same resource.