clips-executive: keep sync map entries on deletes
Do not delete the sync map entry when deleting a fact (either by a
trigger or by a local fact removal). If we delete the sync map entry, we
no longer know when the deletion happened. If we then receive an update
trigger for the same fact ID, we don't know whether that update was
before or after the deletion. Thus, keep the sync map entry, but set the
fact idx to 0 and update the timestamp.