interfaces: do not pack interface structs
Packing interface structs is a bad idea if we access the pointer of a
member of that struct, as the pointer may now be unaligned. We have
never observed memory issues with blackboard interfaces, so packing
seems to be unnecessary anyway.

This effectively reverts 46f1e250c8f6b5938a05a1ad4a347f13285519cb,
except that it removes every packing.

This fixes an address-of-packed member warning that has been added in
GCC9.