gologpp: implement basic skill execution
Construct a skill string using the action mapping defined as part of the
gologpp program and instruct the skiller to execute the skill. Also read
the status from the skiller and pass it back to gologpp, so the
interpreter knows the current state of the action execution.

If the argument of the action is a string, add surrounding quotation
marks. Otherwise, pass it as is.