gologpp: implement action dispatcher to allow other action types
Restructure action execution so we can have a number of different action
executors (e.g., skills, communication). Each executor implements a
function that checks whether the executor is capable of executing a
given activity. The dispatcher iterates over all registered executors
and gives the activity to the first executor that is able to execute the
activity.

The executor is expected to be initialized once (e.g., by opening the
necessary blackboard interfaces) and it should be able to subsequently
execute multiple activities, i.e., we do not create a new executor for
every activity, but only for every type of activity.

Currently, only the SkillerActionExecutor is implemented, which executes
an activity using the skiller. Additional executors can be implemented
by inheriting from the abstract ActionExecutor and registering the
executor with the dispatcher.