gologpp: propagate action preemption to all executors
When preempting an activity, simply pass the activity to all executors,
such that the executor that it is actually running the activity will
stop it.

This also allows multiple concurrently running activities, possibly even
multiple activities per executor, if the executor allows it.

Adapt the function signature of `ActionExecutor::stop` to also accept a
Transition, which is a sub-class of `Grounding<Action>`.