gologpp: turn action executor dispatcher into an aspect
We may want to define multiple executors, e.g., a separate executor for
refbox communication. As these cannot be part of this plugin, provide
access to the dispatcher by creating an aspect that initializes the
dispatcher and passes it to the components using the aspect.

To create a new executor, inherit from the ActionExecutor and also add
the GologppDispatcherAspect to your executor. When initializing the
executor, register it with the dispatcher, so the dispatcher knows what
executor to use for which activity.