gologpp: allow multiple specs in config
Adopt a configuration scheme similar to the scheme known from the CX:
We now have a list of gologpp dirs that contains all the code. In those
directory, each sub-directory is expected to be a spec, i.e., a
particular program for a specific domain. In that directory, the main
program should have the same name as the spec with the extension `.gpp`.

This will allow additional configuration specific to a spec, e.g., we
may only want to watch blackboard interfaces if we have a specific spec
loaded.