gologpp: add action executor for Blackboard messages
This action executor allows sending blackboard messages by executing a
Golog++ action. In order to do this, you need to specify a message
mapping for each action. The mapping must specify the interface_id, the
interface_type, and the message_type that the action should be mapped
to. Furthermore, the action parameters should be the same as the field
names of the message.