format-scripts: do not fail if file does not exist
A non-existing file cannot be in the wrong format, so the
`check-file.sh` script should not fail, but exit with success.

This fixes an issue where the check-format build target passes a deleted
file to the script. Rather than filtering out those files in the caller,
also accept non-existing files and immediately succeed.