gologpp: remove unused `get_executors()` in action dispatcher
The function was used to preempt an activity, as we previously could not
directly preempt an activity, because we were given the
`gologpp::Transition`, not the `gologpp::Activity`. This was changed in
f1857d59424982bcdf282ad15502486ad6f60d0e, which removes the requirement
for `get_executors()`. As there is currently no other use-case for this
member function, simply remove it.