gologpp: fix blackboard ownership handling in skiller executors
The RemoteSkillerActionExecutor creates a new blackboard interface and
thus needs to delete it on destruction. The local SkillerActionExecutor
on the other hand does not own its blackboard and therefore must not
delete it. We can also not delete the blackboard in the destructor of
RemoteSkillerActionExecutor, as the blackboard is still used in the
destructor of SkillerActionExecutor. Thus, add a flag
`blackboard_owner_` which is true only for the
RemoteSkillerActionExecutor and which is used to decide whether to
delete the blackboard in the destructor of SkillerActionExecutor.